move plugin logic to root
