+++ gatsby-browser.js
@@ -1,5 +1,6 @@
 import React from 'react';
 
-import TopLayout from './TopLayout';
+import TopLayout from './src/components/TopLayout';
 
+// eslint-disable-next-line react/jsx-filename-extension
 export const wrapRootElement = ({ element }) => <TopLayout>{element}</TopLayout>;
