--- plugins/gatsby-plugin-top-layout/gatsby-config.js
@@ -1,14 +1,0 @@
-module.exports = {
-  plugins: [
-    `gatsby-transformer-sharp`,
-    `gatsby-plugin-sharp`,
-
-    {
-      resolve: `gatsby-source-filesystem`,
-      options: {
-        name: `markdown-pages`,
-        path: `${__dirname}/../../content`,
-      },
-    },
-  ],
-};
