--- plugins/gatsby-plugin-top-layout/gatsby-ssr.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-import TopLayout from './TopLayout';
-
-// export const wrapRootElement = ({ element }) => <TopLayout>{element}</TopLayout>;
