--- plugins/gatsby-plugin-top-layout/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "gatsby-plugin-top-layout",
-  "dependencies": {
-    "gatsby-plugin-sharp": "^2.6.36",
-    "gatsby-source-filesystem": "^2.3.30",
-    "gatsby-transformer-sharp": "^2.5.15"
-  }
-}
